// src/domain/enums/Department.ts
export enum Department {
    MARKETING = 'Marketing',
    FINANCE = 'Financeiro',
    HR = 'Recursos Humanos',
    TECHNOLOGY = 'Tecnologia',
    ADMINISTRATION = 'Administração',
    COMMERCIAL = 'Comercial',
    OPERATIONS = 'Operações',
    LOGISTICS = 'Logística',
    LEGAL = 'Jurídico',
};